Fix processing of issues which were mentioned in some PR.
Query that we use to retrieve mentions uses GraphQL type matching
to filted out only CrossReferencedEvent out of issue timeline events.
Events that didn't match are still sent as an empty objets. Before
that patch they we filtered away references by checking presense of
the "source" field. But PR mention also mtches CrossReferencedEvent
type, and looks like `{ "id": "...", "source": {}}` -- so it was
passing that check and code after that tried to create an issue out
of `{}`. Fix that by also checking `id` presense in a source.